+++ Views/ErrorWindow.xaml.cs
@@ -1,0 +1,27 @@
+using System.Windows;
+using System.Windows.Input;
+
+namespace CostBIM.Views
+{
+    public partial class ErrorWindow : Window
+    {
+        public ErrorWindow(string message)
+        {
+            InitializeComponent();
+            TxtErrorMessage.Text = message;
+        }
+
+        private void Header_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        {
+            if (e.ButtonState == MouseButtonState.Pressed)
+            {
+                this.DragMove();
+            }
+        }
+
+        private void BtnClose_Click(object sender, RoutedEventArgs e)
+        {
+            this.Close();
+        }
+    }
+}